+++ DSCResources/Browser/Browser.schema.psm1
@@ -97,4 +97,7 @@ Configuration Browser
 
     Import-DscResource -ModuleName xPSDesiredStateConfiguration -ModuleVersion 8.3.0.0
     . "$resourcePath\windows.xRegistry.ps1"
+
+    Import-DscResource -ModuleName PolicyFileEditor -ModuleVersion 3.0.1
+    . "$resourcePath\windows.cAdministrativeTemplateSetting.ps1"
 }