+++ Module/Convert.WebConfigurationPropertyRule/Convert.WebConfigurationPropertyRule.psm1
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 using module .\..\Common\Common.psm1
-using module .\..\Convert.Stig\Convert.Stig.psm1
+using module .\..\Rule\Rule.psm1
 
 $exclude = @($MyInvocation.MyCommand.Name,'Template.*.txt')
 $supportFileList = Get-ChildItem -Path $PSScriptRoot -Exclude $exclude
@@ -28,7 +28,7 @@ Foreach ($supportFile in $supportFileList)
     .PARAMETER Value
         The value the web.config key should be set to
 #>
-Class WebConfigurationPropertyRule : STIG
+Class WebConfigurationPropertyRule : Rule
 {
     [string] $ConfigSection
     [string] $Key
@@ -131,7 +131,7 @@ Class WebConfigurationPropertyRule : STIG
     #>
     static [bool] HasMultipleRules ( [string] $CheckContent )
     {
-        return Test-MultipleWebConfigurationPropertyRule -CheckContent ( [STIG]::SplitCheckContent( $CheckContent ) )
+        return Test-MultipleWebConfigurationPropertyRule -CheckContent ( [Rule]::SplitCheckContent( $CheckContent ) )
     }
 
     <#
@@ -149,7 +149,7 @@ Class WebConfigurationPropertyRule : STIG
     #>
     static [string[]] SplitMultipleRules ( [string] $CheckContent )
     {
-        return ( Split-MultipleWebConfigurationPropertyRule -CheckContent ( [STIG]::SplitCheckContent( $CheckContent ) ) )
+        return ( Split-MultipleWebConfigurationPropertyRule -CheckContent ( [Rule]::SplitCheckContent( $CheckContent ) ) )
     }
 
     #endregion