+++ Tests/Unit/Module/Convert.FileContentRule.tests.ps1
@@ -10,6 +10,7 @@ try
             @{
                 Key          = 'Security.*'
                 Value        = 'MultipleRule'
+                ArchiveFile  = 'MozillaFirefox'
                 CheckContent = 'Open a browser window, type "about:config" in the address bar.
 
                 Verify Preference Name "security.enable_tls" is set to the value "true" and locked.
@@ -18,13 +19,14 @@ try
                 Verify Preference Name "security.tls.version.min" is set to the value "2" and locked.
                 Verify Preference Name "security.tls.version.max" is set to the value "3" and locked.
 
-                Criteria: If the parameters are set incorrectly, then this is a finding. 
+                Criteria: If the parameters are set incorrectly, then this is a finding.
 
                 If the settings are not locked, then this is a finding.'
             }
             @{
                 Key          = 'security.default_personal_cert'
                 Value        = 'Ask Every Time'
+                ArchiveFile  = 'MozillaFirefox'
                 CheckContent = 'Type "about:config" in the browser address bar. Verify  Preference Name "security.default_personal_cert" is set to "Ask Every Time" and is locked to prevent the user from altering.
 
                 Criteria: If the value of "security.default_personal_cert" is set incorrectly or is not locked, then this is a finding.'
@@ -32,14 +34,29 @@ try
             @{
                 Key          = 'plugin.disable_full_page_plugin_for_types'
                 Value        = 'PDF,FDF,XFDF,LSL,LSO,LSS,IQY,RQY,XLK,XLS,XLT,POT,PPS,PPT,DOS,DOT,WKS,BAT,PS,EPS,WCH,WCM,WB1,WB3,RTF,DOC,MDB,MDE,WBK,WB1,WCH,WCM,AD,ADP'
+                ArchiveFile  = 'MozillaFirefox'
                 CheckContent = 'Open a browser window, type "about:config" in the address bar.
 
                 Criteria:  If the "plugin.disable_full_page_plugin_for_types" value is not set to include the following external extensions and not locked, then this is a finding:
 
                 PDF, FDF, XFDF, LSL, LSO, LSS, IQY, RQY, XLK, XLS, XLT, POT PPS, PPT, DOS, DOT, WKS, BAT, PS, EPS, WCH, WCM, WB1, WB3, RTF, DOC, MDB, MDE, WBK, WB1, WCH, WCM, AD, ADP.'
             }
-        )
-        $rule = [FileContentRule]::new( (Get-TestStigRule -ReturnGroupOnly) )
+            @{
+                Key          = 'deployment.security.revocation.check*'
+                Value        = 'MultipleRule'
+                ArchiveFile  = 'OracleJRE'
+                CheckContent = 'If the system is on the SIPRNet, this requirement is NA.
+
+                Navigate to the system-level "deployment.properties" file for JRE.
+
+                The location of the deployment.properties file is defined in <JRE Installation Directory>\Lib\deployment.config
+
+                If the key "deployment.security.revocation.check=ALL_CERTIFICATES" is not present, or is set to "PUBLISHER_ONLY", or "NO_CHECK", this is a finding.
+
+                If the key "deployment.security.revocation.check.locked" is not present, this is a finding.'
+            }
+       )
+       $rule = [FileContentRule]::new( (Get-TestStigRule -ReturnGroupOnly) )
         #endregion
         #region Class Tests
         Describe "$($rule.GetType().Name) Child Class" {
@@ -47,7 +64,7 @@ try
             Context 'Base Class' {
 
                 It "Shoud have a BaseType of STIG" {
-                    $rule.GetType().BaseType.ToString() | Should Be 'STIG'
+                    $rule.GetType().BaseType.ToString() | Should Be 'Rule'
                 }
             }
 
@@ -88,6 +105,7 @@ try
         Describe 'Get-KeyValuePair' {
             foreach ( $rule in $rulesToTest )
             {
+                $global:stigXccdfName = $rule.ArchiveFile
                 if ($rule.Value -ne 'MultipleRule')
                 {
                     It "Should be a Key of '$($rule.Key)'" {
@@ -113,6 +131,8 @@ try
                     $checkContent = Split-TestStrings -CheckContent $rule.CheckContent
                     $keyValuePairs = Get-KeyValuePair -CheckContent $checkContent
                     $result = Test-MultipleFileContentRule -KeyValuePair $checkContent
+
+                    <# 'Enable' property of $rule is missing causing empty string output #>
                     It "Should have Enable equal to: '$($rule.Enable)'" {
 
                         $result | Should Be $true
@@ -124,6 +144,7 @@ try
         #region Function Tests
         Describe "ConvertTo-FileContentRule" {
 
+            $global:stigXccdfName = $rulesToTest[1].ArchiveFile
             $stigRule = Get-TestStigRule -CheckContent $rulesToTest[1].checkContent -ReturnGroupOnly
             $rule = ConvertTo-FileContentRule -StigRule $stigRule